requirements.txt
opencv-python==4.5.5.64
pillow==9.0.1
pytesseract==0.3.9
numpy==1.22.3